requirements.txt
python-telegram-bot
psycopg2-binary
google-generativeai
google-auth
google-auth-oauthlib
google-api-python-client
flask
python-dotenv
Pillow
